chore(suite): rename Data file, export directly
