+++ packages/suite/src/views/wallet/send/components/Options/components/EthereumOptions/index.tsx
@@ -4,7 +4,7 @@ import { Button, Tooltip } from '@trezor/components';
 import { Translation } from 'src/components/suite';
 import { OnOffSwitcher } from 'src/components/wallet';
 import { useSendFormContext } from 'src/hooks/wallet';
-import Data from './components/Data';
+import { Data } from './Data';
 
 const Wrapper = styled.div`
     display: flex;
